@samuel2009 phishing post
Greetings,
We write on behalf of Coinbase to request your assistance based on the Evidence to mitigate harmful cyber activity, specifically customer support fraud at the Internet Presence Location using the inaccurate phone number identified above at the Internet Presence Location.

Similar to illegal robocalls coupled with misleading telephone numbers displayed as caller ID information, the Internet Presence Location is fraudulently masquerading as Coinbase in an attempt to wrongfully gain something of value from consumers. As part of this masquerade, the scammers are displaying an inaccurate phone number for Coinbase or Coinbase support at the Internet Presence Location, which based on the Evidence, is likely being used as part of a scammer phone network either in connection with illegal spoofing or a tech support scam. We are copying relevant authorities to investigate and take appropriate action based on the Evidence, and we ask for your support to take all reasonable and appropriate steps per your acceptable use policy before anyone is harmed.

Please also consider whether you may be able to replace the fraudulent masquerade at the Internet presence location with educational materials to protect consumers against illegal spoofing and fraudulent support scams.

Thank you for your support in this important consumer protection initiative.